### 2025-05-28 #8303 @editorialbot
[PRE REVIEW]: VILLASnode: A Open-Source Real-time Multi-protocol Gateway
**Submitting author:** @al3xa23 (<a href="http://orcid.org/0009-0005-7385-4642">Alexandra  Bach</a>)
**Repository:** https://github.com/VILLASframework/node
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.11.0
