### 2025-05-29 dddef32
feat: add option to read files at once
Signed-off-by: riteshk <ritesh.karki@rwth-aachen.de>
