### 2025-06-02 bc70d82
JOSS: Remove blank lines in metadata
Signed-off-by: Alexandra <alexandra.bach@eonerc.rwth-aachen.de>
