### 2025-06-02 e18ed40
JOSS: Remove blank lines in metadata
Signed-off-by: Alexandra <alexandra.bach@eonerc.rwth-aachen.de>
