+++ .pre-commit-config.yaml
@@ -9,6 +9,7 @@ repos:
     hooks:
       - id: trailing-whitespace
         exclude: \.patch$
+        exclude: ^paper\.md$
       - id: end-of-file-fixer
         exclude: \.(svg|patch)$
       - id: check-ast
@@ -19,6 +20,7 @@ repos:
         exclude: .devcontainer/devcontainer\.json
       - id: check-toml
       - id: check-added-large-files
+        exclude: ^paper\.md$
       - id: pretty-format-json
         # black has its own mind of formatting Jupyter notebooks
         exclude: \.ipynb$|^.devcontainer/devcontainer\.json$|package-lock\.json$
