### 2025-06-02 0b5d883
JOSS: Remove blank lines in metadata
Signed-off-by: Alexandra <alexandra.bach@eonerc.rwth-aachen.de>
