### 2025-06-02 cd32711
JOSS: fix pre-commit
Signed-off-by: Alexandra <alexandra.bach@eonerc.rwth-aachen.de>
