+++ .pre-commit-config.yaml
@@ -8,8 +8,7 @@ repos:
     rev: v5.0.0
     hooks:
       - id: trailing-whitespace
-        exclude: \.patch$
-        exclude: ^paper\.md$
+        exclude: \.patch$|^paper\.md$
       - id: end-of-file-fixer
         exclude: \.(svg|patch)$
       - id: check-ast
@@ -56,4 +55,5 @@ repos:
     rev: "v0.13.0"
     hooks:
       - id: markdownlint
+        exclude: ^paper\.md$
         args: [-r, "~MD013,~MD033,~MD024"]
