### 2025-06-02 e1763a9
JOSS: make JOSS CI compliant
Signed-off-by: Alexandra <alexandra.bach@eonerc.rwth-aachen.de>
