### 2025-06-02 9613eed
JOSS: make JOSS CI compliant
Signed-off-by: Alexandra <alexandra.bach@eonerc.rwth-aachen.de>
