+++ paper.bib
@@ -1,7 +1,3 @@
-% Author: Alexandra Bach <alexandra.bach@eonerc.rwth-aachen.de>
-% SPDX-FileCopyrightText: 2014-2025 Institute for Automation of Complex Power Systems, RWTH Aachen University
-% SPDX-License-Identifier: Apache-2.0
-
 @misc{villasnode_docs,
   author = {Vogel, Steffen and Bach, Alexandra and Potter, Dennis and Stevic, Marija and Pitz, Manuel and Mirz, Markus},
   title = {VILLASnode Documentation},
