### 2025-06-04 #8303 @al3xa23
Hi @danielskatz,

regarding your points:
- Yes, the code is all original to the project.
- generatae pdf is wokring now.
- the references are also found.

Let me know if I need to do something else
