### 2025-06-06 #8303 @danielskatz
@al3xa23 - please don't do this. What I asked you to do is `Please suggest ~5 potential reviewers`, not that you should contact any potential reviewers. And I suggested you use the JOSS reviewer database.
