### 2025-06-09 f9ee298
feat(node-file): implement file reading at file_start
Signed-off-by: Ritesh Karki <ritesh.karki@rwth-aachen.de>
