### 2025-06-12 4d8189a
opal-async: Add new node-type for running VILLASnode as an Asynchronous process in OPAL-RT's RT-LAB
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
