+++ src/opal-async-param.cpp
@@ -1,0 +1,42 @@
+/** Retrieve parameters from OPAL-RTs AsyncApi
+ *
+ * @author Steffen Vogel <post@steffenvogel.de>
+ * @copyright 2014-2022, Institute for Automation of Complex Power Systems, EONERC
+ * @license Apache 2.0
+ *********************************************************************************/
+
+#include <cstring>
+#include <iostream>
+
+#include <AsyncApi.h>
+#include <OpalGenAsyncParamCtrl.h>
+
+int main(int argc, char *argv[]) {
+  Opal_GenAsyncParam_Ctrl params;
+
+  if (argc != 4) {
+    std::cerr << "Usage: AsyncShmemName AsyncShmemSize PrintShmemName"
+              << std::endl;
+    return -1;
+  }
+
+  auto err = OpalOpenAsyncMem(atoi(argv[2]), argv[1]);
+  if (err != EOK) {
+    std::cerr << "Model shared memory not available (" << err << ")"
+              << std::endl;
+    return -1;
+  }
+
+  memset(&params, 0, sizeof(params));
+
+  err = OpalGetAsyncCtrlParameters(&params, sizeof(params));
+  if (err != EOK) {
+    std::cerr << "Could not get controller parameters (" << err << ")"
+              << std::endl;
+    return -1;
+  }
+
+  std::cout << params.StringParam[0] << std::endl;
+
+  return 0;
+}
