+++ lib/empty.cpp
@@ -1,0 +1,7 @@
+/** An empty file
+ *
+ * @author Steffen Vogel <steffen.vogel@opal-rt.com>
+ * @copyright 2014-2022, Institute for Automation of Complex Power Systems, EONERC
+ * @copyright OPAL-RT Germany GmbH
+ * @license Apache 2.0
+ *********************************************************************************/
