+++ lib/nodes/opal_async.cpp
@@ -1,0 +1,416 @@
+/* Node type: OPAL-RT Asynchronous Process (libOpalAsync API)
+ *
+ * Author: Steffen Vogel <steffen.vogel@opal-rt.com>
+ * SPDX-FileCopyrightText: 2023-2024 OPAL-RT Germany GmbH
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include <cstdlib>
+#include <thread>
+#include <vector>
+
+#include <unistd.h>
+
+#include <villas/exceptions.hpp>
+#include <villas/log.hpp>
+#include <villas/nodes/opal_async.hpp>
+#include <villas/utils.hpp>
+
+using namespace villas;
+using namespace villas::node;
+using namespace villas::node::opal;
+using namespace villas::utils;
+
+// This defines the maximum number of signals (doubles) that can be sent
+// or received by any individual OpAsyncSend/Recv blocks in the model. This
+// only applies to the "model <-> asynchronous process" communication.
+#define MAXSENDSIZE 64
+#define MAXRECVSIZE 64
+
+// Private static storage
+static std::string
+    shmemAsyncName; // Shared Memory identifiers and size, provided via argv or config.
+static std::string
+    shmemSystemCtrlName; // Shared Memory identifiers and size, provided via argv or config.
+static size_t
+    shmemAsyncSize; // Shared Memory identifiers and size, provided via argv or config.
+
+static std::vector<unsigned> idsSend,
+    idsRecv; // A list of available send / receive icon IDs.
+static spdlog::sink_ptr sink;
+
+static Opal_GenAsyncParam_Ctrl
+    params; // String and Float parameters, provided by the RT-LAB OpAsyncGenCtrl block.
+
+static std::thread awaiter;
+
+static std::mutex lock; // Protecting state and nodeMap
+static int state;
+static std::map<int, OpalAsyncNode *> nodeMap;
+
+static void waitForModelState(int state) {
+  while (OpalGetAsyncModelState() != state)
+    usleep(1000);
+}
+
+static void waitForAsyncSendRequests() {
+  auto logger = logging.get("node:opal");
+  int ret;
+  unsigned id;
+
+  waitForModelState(STATE_RUN);
+
+  while (true) {
+    ret = OpalWaitForAsyncSendRequest(&id);
+    if (ret != EOK) {
+      state = OpalGetAsyncModelState();
+      if (state == STATE_RESET || state == STATE_STOP)
+        break;
+    }
+
+    try {
+      std::lock_guard<std::mutex> lk(lock);
+      nodeMap.at(id)->markReady();
+    } catch (std::out_of_range &e) {
+      logger->warn("Received send request for unknown id={}", id);
+    }
+  }
+
+  // Model has been stopped or resetted
+  // We stop all OPAL nodes here.
+  for (auto const &x : nodeMap)
+    x.second->stop();
+}
+
+static int dumpGlobal() {
+  auto logger = logging.get("node:opal");
+
+  logger->debug("Controller ID: {}", params.controllerID);
+
+  std::stringstream sss, rss;
+  for (auto i : idsSend)
+    sss << i << " ";
+
+  for (auto i : idsRecv)
+    rss << i << " ";
+
+  logger->debug("Send Blocks: {}", sss.str());
+  logger->debug("Receive Blocks: {}", rss.str());
+
+  logger->debug("Control Block Parameters:");
+  for (int i = 0; i < GENASYNC_NB_FLOAT_PARAM; i++)
[…]

[…]
