+++ lib/nodes/opal_async.cpp
@@ -1,0 +1,523 @@
+/* Node type: OPAL-RT Asynchronous Process (libOpalAsyncApi)
+ *
+ * Author: Steffen Vogel <steffen.vogel@opal-rt.com>
+ * SPDX-FileCopyrightText: 2023-2025 OPAL-RT Germany GmbH
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#include <cstdlib>
+#include <thread>
+#include <vector>
+#include <condition_variable>
+#include <mutex>
+
+#include <unistd.h>
+
+#include <spdlog/details/log_msg_buffer.h>
+#include <spdlog/details/null_mutex.h>
+#include <spdlog/sinks/base_sink.h>
+
+// Define RTLAB before including OpalPrint.h for messages to be sent
+// to the OpalDisplay. Otherwise stdout will be used. */
+#define RTLAB
+
+#include <AsyncApi.h>
+#include <OpalGenAsyncParamCtrl.h>
+#include <OpalPrint.h>
+
+#include <villas/exceptions.hpp>
+#include <villas/format.hpp>
+#include <villas/log.hpp>
+#include <villas/node.hpp>
+#include <villas/popen.hpp>
+#include <villas/sample.hpp>
+#include <villas/utils.hpp>
+
+using namespace villas;
+using namespace villas::node;
+using namespace villas::utils;
+
+// This defines the maximum number of signals (doubles) that can be sent
+// or received by any individual OpAsyncSend/Recv blocks in the model. This
+// only applies to the "model <-> asynchronous process" communication.
+#define MAXSENDSIZE 64
+#define MAXRECVSIZE 64
+
+// Forward declarations
+class OpalAsyncNode;
+
+// Private static storage
+
+// Shared Memory identifiers and size, provided via argv or config.
+static std::string shmemAsyncName;
+
+// Shared Memory identifiers and size, provided via argv or config.
+static std::string shmemSystemCtrlName;
+
+// Shared Memory identifiers and size, provided via argv or config.
+static size_t shmemAsyncSize;
+
+// A list of available send / receive icon IDs.
+static std::vector<unsigned> idsSend, idsRecv;
+
+// An spdlog sink which logs outputs to the OPAL-RT simulation via OpalPrint.
+static spdlog::sink_ptr sink;
+
+// String and Float parameters, provided by the RT-LAB OpAsyncGenCtrl block.
+static Opal_GenAsyncParam_Ctrl params;
+
+static std::thread awaiter;
+
+static std::mutex lock; // Protecting state and nodeMap
+static int state;
+static std::map<int, OpalAsyncNode *> nodeMap;
+
+class LogSink final : public spdlog::sinks::base_sink<std::mutex> {
+private:
+  std::string shmemSystemCtrlName;
+
+public:
+  explicit LogSink(const std::string &shmemName) 
+      : shmemSystemCtrlName(shmemName) {
+    // Enable the OpalPrint function. This prints to the OpalDisplay.
+    int ret =
+        OpalSystemCtrl_Register(const_cast<char *>(shmemSystemCtrlName.c_str()));
+    if (ret != EOK)
+      throw RuntimeError("OpalPrint() access not available ({})", ret);
+  }
+
+  ~LogSink() override {
+    OpalSystemCtrl_UnRegister((char *)shmemSystemCtrlName.c_str());
+  }
+
+protected:
+  void sink_it_(const spdlog::details::log_msg &msg) override {
+    spdlog::memory_buf_t buf;
+
+    buf.clear();
+    formatter_->format(msg, buf);
+
+    char *bufPtr;
+
+    bufPtr = buf.data();
+    bufPtr[buf.size()] = '\0';
+
+    OpalPrint(bufPtr);
+  }
[…]
