### 2025-06-12 97be5d4
feat(node-opal-async): Add new node-type for running VILLASnode as an Asynchronous process in OPAL-RT's RT-LAB
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
