+++ tests/integration/test-config.sh
@@ -13,6 +13,7 @@ CONFIGS=$(find ${SRCDIR}/etc/ -name '*.conf' -o -name '*.json')
 
 for CONFIG in ${CONFIGS}; do
     if [ "$(basename ${CONFIG})" == "opal_orchestra.conf" ] ||
+       [ "$(basename ${CONFIG})" == "opal_async.conf" ] ||
        [ "$(basename ${CONFIG})" == "fpga.conf" ] ||
        [ "$(basename ${CONFIG})" == "paths.conf" ] ||
        [ "$(basename ${CONFIG})" == "tricks.json" ] ||
