+++ doc/openapi/components/schemas/config/nodes/_opal_async.yaml
@@ -1,0 +1,7 @@
+# yaml-language-server: $schema=http://json-schema.org/draft-07/schema
+# SPDX-FileCopyrightText: 2025 Institute for Automation of Complex Power Systems, RWTH Aachen University
+# SPDX-License-Identifier: Apache-2.0
+---
+allOf:
+- $ref: ../node_obj.yaml
+- $ref: opal_async.yaml
