### 2025-06-12 e6d19be
redis: Fix broken log statement
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
