### 2025-06-12 e331b4e
rpm: Add script for build RPM package of VILLASnode
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
