+++ doc/openapi/components/schemas/config/nodes/opal_async.yaml
@@ -1,5 +1,6 @@
 # yaml-language-server: $schema=http://json-schema.org/draft-07/schema
-# SPDX-FileCopyrightText: 2014-2023 Institute for Automation of Complex Power Systems, RWTH Aachen University
+# Author: Steffen Vogel <steffen.vogel@opal-rt.com>
+# SPDX-FileCopyrightText: 2023-2025 OPAL-RT Germany GmbH
 # SPDX-License-Identifier: Apache-2.0
 ---
 allOf:
