### 2025-06-13 #8303 @Puneetha-Ramachandra
Yes, I can do the review.

On Thu, Jun 12, 2025, 2:00 PM The Open Journals editorial robot <
***@***.***> wrote:

> *editorialbot* left a comment (openjournals/joss-reviews#8303)
> <https://github.com/openjournals/joss-reviews/issues/8303#issuecomment-2968141150>
>
> @michmx <https://github.com/michmx> added to the reviewers list!
>
