### 2025-06-13 #8401 @editorialbot
[REVIEW]: VILLASnode: A Open-Source Real-time Multi-protocol Gateway
**Submitting author:** @al3xa23 (<a href="http://orcid.org/0009-0005-7385-4642">Alexandra  Bach</a>)
**Repository:** https://github.com/VILLASframework/node
**Branch with paper.md** (empty if default branch): master
**Version:** v1.0.1
