### 2025-06-16 6b11f13
fix: Formatting, Add author, Use stringstream, proper comments and function name and remove useless code.
Signed-off-by: SystemsPurge <naktiyoussef@proton.me>
