+++ lib/api/requests/metrics.cpp
@@ -1,12 +1,15 @@
-/* The "nodes" API ressource.
+/* The metrics API ressource.
  *
  * Author: Steffen Vogel <post@steffenvogel.de>
+ * Author: Youssef Nakti <naktiyoussef@proton.me>
  * SPDX-FileCopyrightText: 2014-2023 Institute for Automation of Complex Power Systems, RWTH Aachen University
  * SPDX-License-Identifier: Apache-2.0
  */
 
+#include <chrono>
+
 #include <jansson.h>
-#include<chrono>
+
 #include <villas/api/request.hpp>
 #include <villas/api/response.hpp>
 #include <villas/api/session.hpp>
@@ -20,15 +23,6 @@ namespace node {
 namespace api {
 
 class MetricsRequest : public Request {
-private:
-std::unordered_map<Stats::Metric,std::string> metrics_subset = {
-    {Stats::Metric::SMPS_SKIPPED,"skipped"},
-    {Stats::Metric::OWD,"owd"},
-    {Stats::Metric::AGE,"age"},
-    {Stats::Metric::SIGNAL_COUNT,"signalcnt"},
-    {Stats::Metric::RTP_PKTS_LOST,"rtp_pkts_lost"}
-};
-
 public:
   using Request::Request;
 
@@ -37,30 +31,33 @@ public:
       throw InvalidMethod(this);
 
     if (body != nullptr)
-      throw BadRequest("Nodes endpoint does not accept any body data");
-      
-    std::string text_res = "";
+      throw BadRequest("The metrics endpoint does not accept any body data");
+
+    std::stringstream res_stream;
     NodeList node_list = session->getSuperNode()->getNodes();
-    for(Node* node: node_list){
+    for (Node *node : node_list) {
       auto stats = node->getStats();
-      if(!stats)
+      if (!stats)
         continue;
       std::string node_name = node->getNameShort();
-      for(auto& metric:Stats::metrics){
+      for (auto &metric : Stats::metrics) {
         std::string metric_name = metric.second.name;
-        std::replace(metric_name.begin(),metric_name.end(),'.','_');
-        text_res+= stats->getHistogram(metric.first).promFormat(metric_name,node->getNameShort())+"\n";
+        std::replace(metric_name.begin(), metric_name.end(), '.', '_');
+        res_stream << stats->getHistogram(metric.first)
+                          .toPrometheusText(metric_name, node->getNameShort())
+                   << "\n";
       }
     }
-
-    return new Response(session,HTTP_STATUS_OK, "text/plain; charset=UTF-8", Buffer(text_res.c_str(),text_res.size()));
+    auto text_res = res_stream.str();
+    return new Response(session, HTTP_STATUS_OK, "text/plain; charset=UTF-8",
+                        Buffer(text_res.c_str(), text_res.size()));
   }
 };
 
 // Register API request
 static char n[] = "metrics";
 static char r[] = "/metrics";
-static char d[] = "Get stats of all nodes in desired format";
+static char d[] = "Get stats of all nodes in Prometheus metrics format";
 static RequestPlugin<MetricsRequest, n, r, d> p;
 
 } // namespace api
