+++ common/include/villas/hist.hpp
@@ -7,8 +7,9 @@
 
 #pragma once
 
+#include <string>
 #include <vector>
-#include<string>
+
 #include <jansson.h>
 
 #include <villas/log.hpp>
@@ -49,7 +50,6 @@ public:
   // Print ASCII style plot of histogram.
   void plot(Logger logger) const;
 
-  void test();
   // Dump histogram data in Matlab format.
   //
   // @return The string containing the dump. The caller is responsible to free() the buffer.
@@ -61,8 +61,9 @@ public:
   // Write the histogram in JSON format to file \p f.
   int dumpJson(FILE *f) const;
 
-  std::string promFormat(std::string metric_name, std::string node_name) const;
-  
+  std::string toPrometheusText(std::string metric_name,
+                               std::string node_name) const;
+
   // Build a libjansson / JSON object of the histogram.
   json_t *toJson() const;
 
