### 2025-06-16 67fa8ec
feat: add option to read full file contents at once. Fix issues with epoch_mode and eof compatibility with read_mode
Signed-off-by: Ritesh Karki <ritesh.karki@rwth-aachen.de>
