### 2025-06-17 VILLASframework/node#925 @Puneetha-Ramachandra
[joss-reviews] License needs to be updated
openjournals/joss-reviews#8401

There is a LICENSE file in the repository root and also a LICENSE folder under `node/`. Both of them contain template license files without proper year and owner listed.

1. Please consolidate the license file(s)
2. Please update the license templates appropriately (with year, owner and other details filled out)
