### 2025-06-17 57681f0
fix: change method of python api restart node request to POST instead of default GET
Signed-off-by: SystemsPurge <naktiyoussef@proton.me>
