### 2025-06-17 #8401 @stv0g
Hi @Puneetha-Ramachandra, Hi @danielskatz,

We, the authors, have not reached a consensus on the order of the author list.

@al3xa23 [has authored the paper](https://github.com/VILLASframework/node/compare/master...joss) while the software itself has been written by mostly different researchers.

I have written VILLASnode and other software components of the framework as part of my dissertation at the Institute for Automation of Complex Power Systems, lead by Prof. Monti who is also the last author of the paper. Since two years, I have now left the institute and continued developing it as part of my new employment at OPAL-RT.

I propose to alter the author list by using the list of contributors, as ordered above:

- Steffen Vogel
- Niklas Eiling
- Manuel Pitz

In addition I propose to include the following two authors:

- @al3xa23 for authoring the paper, and managing the submission

- Marija Stevic for driving the direction and implementation of the software as part of her dissertation. Her contributions to the software are clearly shown by the past publications about the software which are listed here:

https://villas.fein-aachen.org/docs/publications
