### 2025-06-17 38ed427
feat: add option to read full file contents at once. Remove commented code
Signed-off-by: Ritesh Karki <ritesh.karki@rwth-aachen.de>
