### 2025-06-20 VILLASframework/node#927 @Puneetha-Ramachandra
[joss-review] Installation needs to refresh the ldconfig cache
openjournals/joss-reviews#8401

I was able to build and install the software following instructions at https://villas.fein-aachen.org/docs/node/installation. I also did the optional [installation step](https://villas.fein-aachen.org/docs/node/installation#installation-1). I'm able to run the examples using `<VILLASnode>/build/src/villas-node` and other executables. However when I use `villas node` (i.e from the installed path) I'm running into linking issues.

```
$ villas signal
villas-signal: error while loading shared libraries: libvillas.so.1: cannot open shared object file: No such file or dire
ctory
```

The installation added the libraries to `/usr/local/lib`
```
...
Install the project...
-- Install configuration: "Release"
-- Installing: /usr/local/lib/libvillas-common.so.0.11.0
-- Installing: /usr/local/lib/libvillas-common.so.1
-- Installing: /usr/local/lib/libvillas-common.so
...
```

I had to run `sudo ldconfig` to refresh the cache and things started working. Note this fix was needed for running integration tests as the tests used the binaries from the installed paths.

Perhaps running ldconfig to refresh the cache should be part of the install script?
