### 2025-06-20 VILLASframework/node#928 @Puneetha-Ramachandra
[joss-review] A few of the integration tests are failing
openjournals/joss-reviews#8401

```
 % BUILDDIR=./build/ ./tools/integration-tests.sh | grep FAIL
[FAIL]  api-capabilities (exited with code 127, ran for 1.052004032s)
[FAIL]  api-nodes (exited with code 127, ran for 1.041086657s)
[FAIL]  api-paths (exited with code 127, ran for 1.043206324s)
[FAIL]  hook-shift_ts (exited with code 127, ran for 2.087276998s)
[FAIL]  node-opendss (exited with code 255, ran for 2.036400935s)
[FAIL]  node-stats (exited with code 127, ran for 5.033652901s)
[FAIL]  pipe-loopback-amqp (exited with code 255, ran for .130889109s)
[FAIL]  pipe-loopback-iec61850-9-2 (exited with code 255, ran for .241226687s)
[FAIL]  pipe-loopback-mqtt (exited with code 255, ran for .133253476s)
[FAIL]  test-config (exited with code 255, ran for 3.849568270s)
```
