### 2025-06-20 VILLASframework/node#928 @Puneetha-Ramachandra
I just saw the past run on JOSS branch where the integration test had succeeded -> https://git.rwth-aachen.de/acs/public/villas/node/-/jobs/6334481. Please feel free to close this issue if I'm only supposed to look at the JOSS branch and not the main branch.
