### 2025-06-20 VILLASframework/node#928 @stv0g
The most recent and successful run of the integration tests job on the master branch of VILLASnode can be seen here:

https://git.rwth-aachen.de/acs/public/villas/node/-/jobs/6370858
