### 2025-06-20 VILLASframework/node#927 @stv0g
Thanks @Puneetha-Ramachandra for the report :)

I have adjusted the installation instructions via this PR: https://github.com/VILLASframework/documentation/pull/124/files
