### 2025-06-20 VILLASframework/node#929 @Puneetha-Ramachandra
[joss-review] Paper.md review
openjournals/joss-reviews#8401

Few suggestions -

- [x] 1. Protocol Support Clarity: The paper highlights VILLASnode's role as a multi-protocol gateway. While the documentation (e.g., user guide) indicates support for a wide range of protocols such as AMQP, IEC 61850-9-2, and RTP, the paper itself does not explicitly list examples of these. To better convey the breadth and versatility of the software's communication capabilities, it would be beneficial to include a selection of the key supported protocols directly within the main text.

- [x] 2. Comparison with Existing Tools: The paper effectively articulates VILLASnode's unique positioning, particularly its black-box approach for IP protection in distributed experiments. However, for a comprehensive software overview, it would be valuable to briefly discuss how VILLASnode compares to other commonly used co-simulation frameworks or related software packages (e.g., mosaik, or commercial alternatives in the power systems simulation domain). A concise comparison, highlighting key tradeoffs or advantages of VILLASnode, would enable researchers to make more informed decisions regarding tool selection for their specific applications.

Minor Comments:

- [x] Acknowledgements Duplication: The "ERIGrid2.0" project is listed twice in the Acknowledgements section. Please remove the duplicate entry.
- [x] Affiliation Formatting Consistency: The formatting for multiple affiliations appears inconsistent. For example, affiliation: "1,2" should be standardized to affiliation: "1, 2" for uniformity throughout the author list.
