### 2025-06-20 VILLASframework/node#927 @stv0g
Changes are now live in online documentation: https://villas.fein-aachen.org/docs/node/installation/#installation-1
