### 2025-06-20 a0aec4b
fix(paper): Harmonize style of affiliations
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
