### 2025-06-20 dde2ff3
fix(paper): Remove duplicate ERIGrid acknowledgement
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
