### 2025-06-20 eee65bd
fix(paper): Use LaTeX mode for referencing symbols in figure
Signed-off-by: Steffen Vogel <post@steffenvogel.de>
