### 2025-06-20 df86077
fix(paper): Move JOSS paper into subdirectory
Signed-off-by: Steffen Vogel <post@steffenvogel.de>
