### 2025-06-20 3f7c80a
fix(paper): Add a paragraph highlighting which data exchange protocols are supported
Signed-off-by: Steffen Vogel <post@steffenvogel.de>
