### 2025-06-20 fd6b32b
fix(paper): Add a paragraph highlighting which data exchange protocols are supported
Signed-off-by: Steffen Vogel <post@steffenvogel.de>
