### 2025-06-20 VILLASframework/node#929 @stv0g
In https://github.com/VILLASframework/node/commit/fd6b32bcc1e66b451d536d9e3983c66b2b60fd59, I added a paragraph to the paper describing which data exchange protocols are supported.
