### 2025-06-20 5e9d5a0
fix(paper): Update date
Signed-off-by: Steffen Vogel <post@steffenvogel.de>
