### 2025-06-20 9f9d43d
fix(paper): Fix REUSE exceptions
Signed-off-by: Steffen Vogel <post@steffenvogel.de>
