+++ REUSE.toml
@@ -24,9 +24,9 @@ path = [
     "etc/labs/lab3.pcap",
     "tests/valgrind.supp",
     "packaging/archlinux/villas-node.install",
-    "VILLASnode_paths.svg",
-    "paper.md",
-    "paper.bib",
+    "paper/figures/**",
+    "paper/paper.md",
+    "paper/paper.bib",
 ]
 precedence = "aggregate"
 SPDX-FileCopyrightText = "2018-2023, Institute for Automation of Complex Power Systems, RWTH Aachen University"
