### 2025-06-20 24d2dbe
fix(paper): Fix markdownlint errors
Signed-off-by: Steffen Vogel <post@steffenvogel.de>
