### 2025-06-20 ecb4cd5
feat(paper): Add badge to README file
Signed-off-by: Steffen Vogel <post@steffenvogel.de>
