### 2025-06-20 9594ae3
feat: add option to read full file contents at once. fix: refine code symantics.
Signed-off-by: Ritesh Karki <ritesh.karki@rwth-aachen.de>
