### 2025-06-23 f809870
fix(paper): Proposal for new author list
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
