### 2025-06-24 84c0a8d
feat(api): Add a new metrics endpoint for exporting statistics to Prometheus
Co-authored-by: SystemsPurge <naktiyoussef@proton.me>
Signed-off-by: SystemsPurge <naktiyoussef@proton.me>
Co-authored-by: Steffen Vogel <steffen.vogel@opal-rt.com>
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
