### 2025-06-24 19b2cea
feat(api): Add a new metrics endpoint for exporting statistics to Prometheus
Co-authored-by: SystemsPurge <naktiyoussef@proton.me>
Signed-off-by: SystemsPurge <naktiyoussef@proton.me>
Co-authored-by: Steffen Vogel <steffen.vogel@opal-rt.com>
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
