+++ tests/integration/api-metrics.sh
@@ -1,0 +1,158 @@
+#!/usr/bin/env bash
+#
+# Integration test for remote API
+#
+# Author: Steffen Vogel <steffen.vogel@opal-rt.com>
+# SPDX-FileCopyrightText: 2025 OPAL-RT Germany GmbH
+# SPDX-License-Identifier: Apache-2.0
+
+set -e
+
+DIR=$(mktemp -d)
+pushd ${DIR}
+
+function finish {
+    popd
+    rm -rf ${DIR}
+}
+trap finish EXIT
+
+cat > config.json <<EOF
+{
+    "http": {
+        "port": 8080
+    },
+    "nodes": {
+        "node1": {
+            "type": "signal",
+
+            "signal": "sine",
+            "limit": 10,
+            "values": 1,
+
+            "in": {
+                "hooks": [
+                    {
+                        "type": "stats",
+                        "warmup": 0,
+                        "buckets": 5
+                    }
+                ]
+            }
+        }
+    },
+    "paths": [
+        {
+            "in": "node1"
+        }
+    ]
+}
+EOF
+
+cat <<EOF > expected
+#TYPE HISTOGRAM rtp_jitter
+rtp_jitter {node="node1" le="0"} 0
+rtp_jitter {node="node1" le="0"} 0
+rtp_jitter {node="node1" le="0"} 0
+rtp_jitter {node="node1" le="0"} 0
+rtp_jitter {node="node1" le="0"} 0
+rtp_jitter {node="node1" le="+Inf"} 0
+rtp_jitter_count  {node="node1"} 0
+
+#TYPE HISTOGRAM rtp_pkts_lost
+rtp_pkts_lost {node="node1" le="0"} 0
+rtp_pkts_lost {node="node1" le="0"} 0
+rtp_pkts_lost {node="node1" le="0"} 0
+rtp_pkts_lost {node="node1" le="0"} 0
+rtp_pkts_lost {node="node1" le="0"} 0
+rtp_pkts_lost {node="node1" le="+Inf"} 0
+rtp_pkts_lost_count  {node="node1"} 0
+
+#TYPE HISTOGRAM rtp_loss_fraction
+rtp_loss_fraction {node="node1" le="0"} 0
+rtp_loss_fraction {node="node1" le="0"} 0
+rtp_loss_fraction {node="node1" le="0"} 0
+rtp_loss_fraction {node="node1" le="0"} 0
+rtp_loss_fraction {node="node1" le="0"} 0
+rtp_loss_fraction {node="node1" le="+Inf"} 0
+rtp_loss_fraction_count  {node="node1"} 0
+
+#TYPE HISTOGRAM signal_cnt
+signal_cnt {node="node1" le="-8"} 0
+signal_cnt {node="node1" le="-4"} 0
+signal_cnt {node="node1" le="0"} 0
+signal_cnt {node="node1" le="4"} 9
+signal_cnt {node="node1" le="8"} 9
+signal_cnt {node="node1" le="+Inf"} 10
+signal_cnt_count  {node="node1"} 10
+
+#TYPE HISTOGRAM age
+age {node="node1" le="0"} 0
+age {node="node1" le="0"} 0
+age {node="node1" le="0"} 0
+age {node="node1" le="0"} 0
+age {node="node1" le="0"} 0
+age {node="node1" le="+Inf"} 0
+age_count  {node="node1"} 0
+
+#TYPE HISTOGRAM owd
+owd {node="node1" le="-8"} 0
+owd {node="node1" le="-4"} 0
+owd {node="node1" le="0"} 0
+owd {node="node1" le="4"} 8
+owd {node="node1" le="8"} 8
+owd {node="node1" le="+Inf"} 9
+owd_count  {node="node1"} 9
+
[…]
