+++ tests/integration/node-stats.sh
@@ -23,41 +23,41 @@ cat > config.json <<EOF
 {
     "nodes": {
         "stats_1": {
-             "type": "stats",
-             "node": "signal_1",
-             "rate": 10.0,
-             "in": {
-             	"signals": [
-             		{ "name": "gap",   "stats": "signal_1.gap_sent.mean" },
-             		{ "name": "total", "stats": "signal_1.owd.total" }
-             	]
-             }
+            "type": "stats",
+            "node": "signal_1",
+            "rate": 10.0,
+            "in": {
+                "signals": [
+                    { "name": "gap",   "stats": "signal_1.gap_sent.mean" },
+                    { "name": "total", "stats": "signal_1.owd.total" }
+                ]
+            }
         },
         "signal_1": {
-             "type": "signal",
-             "limit": 100,
-             "signal": "sine",
-             "rate": ${RATE},
-             "in": {
-             	"hooks": [
-             		{ "type": "stats", "verbose": true }
-             	]
-             }
+            "type": "signal",
+            "limit": 100,
+            "signal": "sine",
+            "rate": ${RATE},
+            "in": {
+                "hooks": [
+                    { "type": "stats", "verbose": true }
+                ]
+            }
         },
         "stats_log_1": {
-             "type": "file",
-             "format": "json",
+            "type": "file",
+            "format": "json",
 
-             "uri": "stats.json"
+            "uri": "stats.json"
         }
     },
     "paths": [
         {
-             "in": "signal_1"
+            "in": "signal_1"
         },
         {
-             "in": "stats_1",
-             "out": "stats_log_1"
+            "in": "stats_1",
+            "out": "stats_log_1"
         }
     ]
 }
