+++ tests/integration/node-multiple-sources.sh
@@ -25,40 +25,40 @@ trap finish EXIT
 
 cat > config.json <<EOF
 {
-        "nodes": {
-             	"sig_1": {
-             			"type": "signal",
-             			"values": 1,
-             			"signal": "counter",
-             			"offset": 100,
-             			"limit": 10,
-             			"realtime": false
-             	},
-             	"file_1": {
-             			"type": "file",
-             			"uri": "output1.dat"
-             	},
-             	"file_2": {
-             			"type": "file",
-             			"uri": "output2.dat"
-             	}
+    "nodes": {
+        "sig_1": {
+            "type": "signal",
+            "values": 1,
+            "signal": "counter",
+            "offset": 100,
+            "limit": 10,
+            "realtime": false
         },
-        "paths": [
-             {
-             	"in": "sig_1",
-             	"out": "file_1"
-             },
-             {
-             	"in": "sig_1",
-             	"out": "file_2"
-             },
-             {
-             	"in": "sig_1"
-             },
-             {
-             	"in": "sig_1"
-             }
-        ]
+        "file_1": {
+            "type": "file",
+            "uri": "output1.dat"
+        },
+        "file_2": {
+            "type": "file",
+            "uri": "output2.dat"
+        }
+    },
+    "paths": [
+        {
+            "in": "sig_1",
+            "out": "file_1"
+        },
+        {
+            "in": "sig_1",
+            "out": "file_2"
+        },
+        {
+            "in": "sig_1"
+        },
+        {
+            "in": "sig_1"
+        }
+    ]
 }
 EOF
 
