### 2025-06-24 1e0b3d4
fix(tests): Fix indentation of test configs
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
