+++ tests/integration/node-rate.sh
@@ -30,31 +30,31 @@ cat > config.json <<EOF
 {
     "nodes": {
         "node_1": {
-             "type": "socket",
+            "type": "socket",
 
-             "in": {
-             	"address": ":12000",
+            "in": {
+                "address": ":12000",
 
-             	"signals": [
-             		{ "name": "sig1", "init": 1234.0 },
-             		{ "name": "sig2", "init": 5678.0 }
-             	]
-             },
-             "out": {
-             	"address": "127.0.0.1:12000"
-             }
+                "signals": [
+                    { "name": "sig1", "init": 1234.0 },
+                    { "name": "sig2", "init": 5678.0 }
+                ]
+            },
+            "out": {
+                "address": "127.0.0.1:12000"
+            }
         },
         "file_1": {
-             "type": "file",
-             "uri": "output.dat"
+            "type": "file",
+            "uri": "output.dat"
         }
     },
     "paths": [
         {
-             "in": "node_1",
-             "out": "file_1",
-             "mode": "all",
-             "rate": 10
+            "in": "node_1",
+            "out": "file_1",
+            "mode": "all",
+            "rate": 10
         }
     ]
 }
