### 2025-06-26 fbb4555
fix: Portability of vstrcatf and vstrf Specifiying the "va" parameter as reference to ensure reproducable behavior on arm and x86.
Signed-off-by: Pascal Bauer <pascal.bauer@rwth-aachen.de>
