### 2025-06-26 844a1c1
fix: vstrcatf wil leave va_list parameter valid after use vstrcatf now copies the va_list parameter to prevent it beeing undetermined after call to vasprintf. See full desciption in #678.
Signed-off-by: Pascal Bauer <pascal.bauer@rwth-aachen.de>
