### 2025-06-26 22255af
add: miob.conf ecample for platform based fpgas Example can use hw loopback.
Signed-off-by: Pascal Bauer <pascal.bauer@rwth-aachen.de>

Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
