+++ tests/integration/test-config.sh
@@ -15,6 +15,7 @@ for CONFIG in ${CONFIGS}; do
     if [ "$(basename ${CONFIG})" == "opal_orchestra.conf" ] ||
        [ "$(basename ${CONFIG})" == "opal_async.conf" ] ||
        [ "$(basename ${CONFIG})" == "fpga.conf" ] ||
+       [ "$(basename ${CONFIG})" == "miob.conf" ] ||
        [ "$(basename ${CONFIG})" == "paths.conf" ] ||
        [ "$(basename ${CONFIG})" == "tricks.json" ] ||
        [ "$(basename ${CONFIG})" == "tricks.conf" ] ||
