### 2025-06-26 1398c5c
add: miob.conf ecample for platform based fpgas Example can use hw loopback.
Signed-off-by: Pascal Bauer <pascal.bauer@rwth-aachen.de>
