+++ tests/integration/test-config.sh
@@ -14,6 +14,7 @@ CONFIGS=$(find ${SRCDIR}/etc/ -name '*.conf' -o -name '*.json')
 for CONFIG in ${CONFIGS}; do
     if [ "$(basename ${CONFIG})" == "opal.conf" ] ||
        [ "$(basename ${CONFIG})" == "fpga.conf" ] ||
+       [ "$(basename ${CONFIG})" == "miob.conf" ] ||
        [ "$(basename ${CONFIG})" == "paths.conf" ] ||
        [ "$(basename ${CONFIG})" == "tricks.json" ] ||
        [ "$(basename ${CONFIG})" == "tricks.conf" ] ||
