### 2025-06-27 854316d
fix(table): Improve table printing for ARM architecture
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
