### 2025-06-27 1fc2ca1
fix(table): Improve table printing for ARM architecture
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
