### 2025-06-27 133574f
fix(table): Improve table printing for ARM architecture
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
