+++ common/include/villas/table.hpp
@@ -27,11 +27,12 @@ protected:
   int _width; // The real width of this column. Calculated by Table::resize().
 
   int width; // Width of the column.
+  int precision; // Precision of the column, used for floating point values.
 
 public:
-  TableColumn(int w, enum Alignment a, const std::string &t,
+  TableColumn(int w, int p, enum Alignment a, const std::string &t,
               const std::string &f, const std::string &u = "")
-      : _width(0), width(w), title(t), format(f), unit(u), align(a) {}
+      : _width(0), width(w), precision(p), title(t), format(f), unit(u), align(a) {}
 
   std::string title;  // The title as shown in the table header.
   std::string format; // The format which is used to print the table rows.
@@ -46,10 +47,11 @@ class Table {
 
 protected:
   int resize(int w);
+  void updateRowFormat();
 
   int width;
-
   std::vector<TableColumn> columns;
+  std::string rowFormat;
 
   Logger logger;
 
