### 2025-06-27 5a6b288
fix(stats,table): Improve table rendering
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
