### 2025-06-27 cd87bf4
fix(stats,table): Improve table rendering
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
