### 2025-06-27 ab45770
fix(table): Avoid allocations
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
