### 2025-06-27 604f80e
fix(table): Avoid allocations
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
