### 2025-06-27 e9c0464
fix(table): Avoid allocations
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
