### 2025-07-01 VILLASframework/node#936 @michmx
[joss-review] Update LICENSE file
https://github.com/openjournals/joss-reviews/issues/8401

- [LICENSE](https://github.com/VILLASframework/node/blob/master/LICENSE) file: Must be updated with the template as described at the bottom of https://opensource.org/license/apache-2-0

```
   Copyright [yyyy] [name of copyright owner]

   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
   You may obtain a copy of the License at

     http://www.apache.org/licenses/LICENSE-2.0

   Unless required by applicable law or agreed to in writing, software
   distributed under the License is distributed on an "AS IS" BASIS,
   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
   See the License for the specific language governing permissions and
   limitations under the License.
```
