### 2025-07-02 VILLASframework/node#936 @stv0g
Hi @michmx,

your fellow reviewer raised a similar issue here: https://github.com/VILLASframework/node/issues/925

I've tried to explain our reasoning in this linked issue.
