### 2025-07-02 VILLASframework/node#936 @michmx
Hi @stv0g,  understood. Then I suggest to update the [README.md](https://github.com/VILLASframework/node/blob/master/README.md) as currently it states

```
This project is released under the terms of the [Apache 2.0 license](https://github.com/VILLASframework/node/blob/master/LICENSE).
```

Adding a similar description as what is explained at https://github.com/VILLASframework/node/issues/925 would avoid the confusion raised by both reviewers.
