### 2025-07-03 VILLASframework/node#937 @michmx
[joss-review] Getting Started page
https://github.com/openjournals/joss-reviews/issues/8401

The Getting Started guide works great. Just a couple of comments to help users with a smoother quick start.

1) The preparation on top of the [Getting Started](https://villas.fein-aachen.org/docs/node/guides/getting_started) page asks to mount a volume where the configs file are
```
--volume C:\path\to\configs:/configs
```

While, only until reaching the end of the page one can find the note:
```
All configurations files are available here: https://github.com/VILLASframework/node/tree/master/etc/labs
```

I suggest to move this note higher in the page, at the level of 'Preparation'.

2) For the purpose of a quick start, it looks that the config files are inside the container under `/villas/etc/labs`. Is this useful to follow the Lab assignments? If yes, suggestion to mention it.
