### 2025-07-03 VILLASframework/node#938 @michmx
[joss-review] Paper.md review
https://github.com/openjournals/joss-reviews/issues/8401

The paper is well written and can be followed by a non-specialist audience.  I just add a few minor suggestions:

- [ ] It may be useful to add in the summary a sentence saying explicitly that the tool was originally designed for simulation of electrical networks, but can be used without restrictions for simulation in other domains, similar as stated in [the web documentation](https://villas.fein-aachen.org/docs/#villasframework-1).

- [ ] It would be ilustrative to provide at least an example of a project using VILLASnode, with a reference to the paper/proceeding if available

- [ ] Python software tools are mentioned when talking about the availability of an API to control VILLASnode. As it is [exposed as a HTTP/REST API](https://github.com/VILLASframework/node/blob/master/doc/openapi/openapi.yaml), in principle the interface can be language-agnostic. I suggest to rephase or remove the sentence "This is useful if Python software tools are to be integrated...", as client applications written in other languages are discussed just after.
