### 2025-07-08 5f726f7
fix(webrtc): Move protocol callback logic to static handler
This fixes a bug where libwebsockets would call the protocol
handler with a NULL `user` pointer on initialization. This caused the
`SignalingClient::protocolCallback` non-static function to be invoked on a
null pointer instance.
