### 2025-07-08 893caa4
fix(common): Add virtual destructor for `Tool` class
Signed-off-by: Philipp Jungkamp <philipp.jungkamp@rwth-aachen.de>
