### 2025-07-08 5514ea8
fix(formats/protobuf): Don't mix new and free
Signed-off-by: Philipp Jungkamp <philipp.jungkamp@rwth-aachen.de>
