### 2025-07-09 0bb9ca2
fix(doc): Update description of VILLASnode's licensing
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
