### 2025-07-09 dbd4f5a
fix(lint): Format code with clang-format
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
