### 2025-07-09 16d7c59
fix(lint): Format code with clang-format
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
