### 2025-07-09 bd505a3
fix(lint): Format code with clang-format
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
