+++ common/include/villas/table.hpp
@@ -28,13 +28,14 @@ public:
 protected:
   int _width; // The real width of this column. Calculated by Table::resize().
 
-  int width; // Width of the column.
+  int width;     // Width of the column.
   int precision; // Precision of the column, used for floating point values.
 
 public:
   TableColumn(int w, int p, enum Alignment a, const std::string &t,
               const std::string &f, const std::string &u = "")
-      : _width(0), width(w), precision(p), title(t), format(f), unit(u), align(a) {}
+      : _width(0), width(w), precision(p), title(t), format(f), unit(u),
+        align(a) {}
 
   std::string title;  // The title as shown in the table header.
   std::string format; // The format which is used to print the table rows.
@@ -65,8 +66,7 @@ public:
   void header();
 
   // Print table rows.
-  template <class... Args>
-  void row(const Args&... args) {
+  template <class... Args> void row(const Args &...args) {
     auto logWidth = Log::getWidth();
     if (width != logWidth) {
       resize(logWidth);
