### 2025-07-09 f6fb690
fix(lint): Remove trailing whitespace
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
