### 2025-07-09 VILLASframework/node#936 @stv0g
Dear @michmx,

we have merged a change which clarifies the per-file licensing and preferred choice of the Apache 2.0 software license in #939
