### 2025-07-09 1d09c7d
docs: add reference to "The VILLASframework Authors" to files in LICENSE folder
Signed-off-by: Niklas Eiling <niklas@eil.ing>
