+++ LICENSES/BSD-3-Clause.txt
@@ -1,4 +1,4 @@
-Copyright (c) 2017 Institute for Automation of Complex Power Systems, RWTH Aachen University.
+Copyright (c) 2014-2025 The VILLASframework Authors.
 
 Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
 
