### 2025-07-09 283108c
use "The VILLASframwork Authors" in shell output and in WebRTC example
Signed-off-by: Niklas Eiling <niklas@eil.ing>
