### 2025-07-09 77c4b44
Use "The VILLASframework Authors" also in CMakeLists Copyright notice and in README.md
Signed-off-by: Niklas Eiling <niklas@eil.ing>
