+++ README.md
@@ -56,11 +56,7 @@ We kindly ask all academic publications employing components of VILLASframework 
 
 For other licensing options please consult [Prof. Antonello Monti](mailto:amonti@eonerc.rwth-aachen.de).
 
-- SPDX-FileCopyrightText: 2014-2025 Institute for Automation of Complex Power Systems, RWTH Aachen University
-- SPDX-FileCopyrightText: 2023-2025 OPAL-RT Germany GmbH
-- SPDX-FileCopyrightText: 2022-2025 Niklas Eiling <niklas.eiling@eonerc.rwth-aachen.de>
-- SPDX-FileCopyrightText: 2018-2025 Steffen Vogel <post@steffenvogel.de>
-- SPDX-FileCopyrightText: 2018 Daniel Krebs <dkrebs@eonerc.rwth-aachen.de>
+- SPDX-FileCopyrightText: 2014-2025 The VILLASframework Authors
 - SPDX-License-Identifier: Apache-2.0
 
 ## Contact
@@ -68,7 +64,7 @@ For other licensing options please consult [Prof. Antonello Monti](mailto:amonti
 [![EONERC ACS Logo](doc/pictures/eonerc_logo.png)](http://www.acs.eonerc.rwth-aachen.de)
 
 - Steffen Vogel <post@steffenvogel.de>
-- Niklas Eiling <niklas.eiling@eonerc.rwth-aachen.de>
+- Niklas Eiling <niklas@eil.ing>
 - Felix Wege <fwege@eonerc.rwth-aachen.de>
 - Alexandra Bach <alexandra.bach@eonerc.rwth-aachen.de>
 
