### 2025-07-10 9771f9c
fix(paper): address comments
Signed-off-by: Alexandra <alexandra.bach@eonerc.rwth-aachen.de>
