+++ paper/paper.bib
@@ -1,3 +1,34 @@
+@article{hardy_helics_2024,
+	title = {{HELICS}: A Co-Simulation Framework for Scalable Multi-Domain Modeling and Analysis},
+	volume = {12},
+	issn = {2169-3536},
+	url = {https://ieeexplore.ieee.org/document/10424422},
+	doi = {10.1109/ACCESS.2024.3363615},
+	shorttitle = {{HELICS}},
+	abstract = {As both the generation resources and load types have changed and grown over the past few decades, there is a growing need for analysis that spans traditional simulation boundaries; for example, evaluating the impact of distribution-level assets (e.g. rooftop solar, {EV} chargers) on bulk-power system operation. Co-simulation is a technique that allows simulators to trade information during run-time, effectively creating larger and more complex models. {HELICS} is a co-simulation platform that has been developed to enable these kinds of power system analysis, incorporating tools from a variety of domains including the electrical power grid, natural gas, transportation, and communications. This paper summarizes the technical design of {HELICS}, describes how tools can be integrated into the platform, and reviews a number of analyses that have been performed using {HELICS}. A short video summary of this paper can be found at https://youtu.be/{BIUiR}\_K87Wc.},
+	pages = {24325--24347},
+	journaltitle = {{IEEE} Access},
+	author = {Hardy, Trevor D. and Palmintier, Bryan and Top, Philip L. and Krishnamurthy, Dheepak and Fuller, Jason C.},
+	urldate = {2025-07-10},
+	date = {2024},
+	keywords = {co-simulation, energy system analysis, Government, {HELICS}, multi-domain analysis, multi-energy analysis, natural gas, Natural gas, Power system analysis computing, Power system dynamics, power system simulation, Power system stability, Renewable energy sources, Scalability, Simulation, Transactive energy, transportation},
+}
+
+@inproceedings{rohjans_mosaik_2013,
+	title = {mosaik - A modular platform for the evaluation of agent-based Smart Grid control},
+	url = {https://ieeexplore.ieee.org/document/6695486},
+	doi = {10.1109/ISGTEurope.2013.6695486},
+	abstract = {Smart Grids rely on the use of {ICT} for managing large numbers of active components and sensors to keep demand and generation of electricity at equilibrium while operating multiple resources within their operational limits. Due to the distributed nature of these resources, their heterogeneity as well as their sheer number, is a challenging task. Control strategies as well as novel paradigms need to be developed and thoroughly evaluated through extensive simulations. In order to yield scientifically sound and reliable results, these simulations have to rely on valid and (ideally) established models, e.g., from industry. Therefore, it is desirable to reuse these models as often as possible by combining them into new, potentially large-scale test scenarios. The introduced mosaik framework presents a flexible architecture as well as a powerful modeling and specification language to automate the process of composing existing models and simulation platforms into large-scale simulation scenarios.},
+	eventtitle = {{IEEE} {PES} {ISGT} Europe 2013},
+	pages = {1--5},
+	booktitle = {{IEEE} {PES} {ISGT} Europe 2013},
+	author = {Rohjans, S. and Lehnhoff, S. and Schütte, S. and Scherfke, S. and Hussain, S.},
+	urldate = {2025-07-10},
+	date = {2013-10},
+	note = {{ISSN}: 2165-4824},
+	keywords = {Biological system modeling, Composition, Computational modeling, {ICT}, Integration, Load modeling, Mathematical model, Modeling, Simulation, Smart Grid Control, Smart grids, Topology},
+}
+
 @misc{villasnode_docs,
   author = {Vogel, Steffen and Bach, Alexandra and Potter, Dennis and Stevic, Marija and Pitz, Manuel and Mirz, Markus},
   title = {VILLASnode Documentation},
