### 2025-07-16 225e776
fix: remove unecessary } in format string in table.cpp
Signed-off-by: Niklas Eiling <niklas@eil.ing>
