### 2025-07-16 8bc146a
fix: improve compatiblity of villas::utils::Table with newer C++ standards
Co-authored-by: Philipp Jungkamp <philipp.jungkamp@rwth-aachen.de>
Signed-off-by: Niklas Eiling <niklas@eil.ing>
