### 2025-07-17 d9d12de
Moved clients/python-binding/ to python/binding/
Signed-off-by: Kevin Vu te Laar <vu.te@rwth-aachen.de>
