### 2025-07-17 24a15ae
feat(config): Add new helper function for parsing human readable durations
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
