### 2025-07-17 ebea969
feat(config): Add new helper function for parsing human readable durations
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
