### 2025-07-17 c6099d4
feat(config): Add new helper function for parsing human readable durations
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
