### 2025-07-17 a146ecb
python-binding unit test split, socket fixes
Unit test portion split.
Previously commented-out tests are now active.
Fixes to the socket node resolved previous issues during testing.
These arised when C-API functions were called through the bindings in likely unintended ways,
causing undefined behavior (unexpected return values), segfaults or other memory related crashes.

Fixes:
- Socket descriptor (`sd`) is now initialized with -1
- `in` and `out` buffers are initialized with `nullptr`
- Defensive deallocation and cleanup of buffers:
  * Check buffers before freeing
  * Zero memory before fr
[…]
