+++ tests/unit/python_binding.py
@@ -20,162 +20,147 @@ class SimpleWrapperTests(unittest.TestCase):
         except Exception as e:
             self.fail(f"new_node err: {e}")
 
-    def test_activity_changes(self):
+    def tearDown(self):
         try:
-            vn.node_check(self.test_node)
-            vn.node_prepare(self.test_node)
-            # starting twice
-            self.assertEqual(0, vn.node_start(self.test_node))
-            # with self.assertRaises((AssertionError, RuntimeError)):
-            #     vn.node_start(self.test_node)
-
-            # check if the node is running
-            self.assertTrue(vn.node_is_enabled(self.test_node))
-
-            # pausing twice
-            self.assertEqual(0, vn.node_pause(self.test_node))
-            self.assertEqual(-1, vn.node_pause(self.test_node))
-
-            # resuming
-            self.assertEqual(0, vn.node_resume(self.test_node))
-
-            # stopping twice
-            self.assertEqual(0, vn.node_stop(self.test_node))
-            # self.assertEqual(-1, vn.node_stop(self.test_node))
-
-            # # restarting
-            # vn.node_restart(self.test_node)
-
-            # check if everything still works after restarting
-            # vn.node_pause(self.test_node)
-            # vn.node_resume(self.test_node)
-            # vn.node_stop(self.test_node)
-
-            # terminating the node
+            vn.node_stop(self.test_node)
             vn.node_destroy(self.test_node)
         except Exception as e:
-            self.fail(f" err: {e}")
+            self.fail(f"node cleanup error: {e}")
 
-    def test_details(self):
+    def test_start(self):
+        try:
+            self.assertEqual(0, vn.node_start(self.test_node))
+            # socket node will cause a RuntimeError
+            # the behavior is not consistent for each node
+            # with self.assertRaises((AssertionError, RuntimeError)):
+            #     vn.node_start(self.test_node)
+        except Exception as e:
+            self.fail(f"err: {e}")
+
+    def test_new(self):
+        try:
+            node_uuid = str(uuid.uuid4())
+            node_config = json.dumps(test_node_config, indent=2)
+            node = vn.node_new(node_uuid, node_config)
+            self.assertIsNotNone(node)
+        except Exception as e:
+            self.fail(f"err: {e}")
+
+    def test_check(self):
+        try:
+            vn.node_check(self.test_node)
+        except Exception as e:
+            self.fail(f"err: {e}")
+
+    def test_prepare(self):
+        try:
+            vn.node_prepare(self.test_node)
+        except Exception as e:
+            self.fail(f"err: {e}")
+
+    def test_is_enabled(self):
+        try:
+            self.assertTrue(vn.node_is_enabled(self.test_node))
+        except Exception as e:
+            self.fail(f"err: {e}")
+
+    def test_pause(self):
+        try:
+            self.assertEqual(-1, vn.node_pause(self.test_node))
+            self.assertEqual(-1, vn.node_pause(self.test_node))
+        except Exception as e:
+            self.fail(f"err: {e}")
+
+    def test_resume(self):
+        try:
+            self.assertEqual(0, vn.node_resume(self.test_node))
+        except Exception as e:
+            self.fail(f"err: {e}")
+
+    def test_stop(self):
+        try:
+            self.assertEqual(0, vn.node_stop(self.test_node))
+            self.assertEqual(0, vn.node_stop(self.test_node))
+            vn.node_restart(self.test_node)
+        except Exception as e:
+            self.fail(f"err: {e}")
+
+    def test_restart(self):
+        try:
+            self.assertEqual(0, vn.node_restart(self.test_node))
+            self.assertEqual(0, vn.node_restart(self.test_node))
+        except Exception as e:
+            self.fail(f"err: {e}")
+
[…]
