### 2025-07-17 b9a0f9f
fix(node-opal-orchestra): Avoid line-breaks my relocate comments
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
