### 2025-07-17 21aed88
feat(deps): Add little helper function to check for availability of commands
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
