### 2025-07-17 0af1fc8
feat(deps): Install componentorchestra library via dpkg
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
