### 2025-07-17 d7d2e08
feat(deps): Install componentorchestra library via dpkg
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
