### 2025-07-22 9fa8f4b
tests: Add R-GOOSE integration test
Signed-off-by: Philipp Jungkamp <philipp.jungkamp@rwth-aachen.de>
