### 2025-07-24 2d79441
fix(fpga): improve error message when we we are unable to rewrite the BAR
Signed-off-by: Niklas Eiling <niklas@eil.ing>
