+++ common/lib/kernel/vfio_container.cpp
@@ -214,8 +214,9 @@ std::shared_ptr<Device> Container::attachDevice(devices::PciDevice &pdev) {
         e.what() +
         std::string(
             "\nBAR of device is in inconsistent state. Rewriting the BAR "
-            "failed. Please remove, rescan and reset the device and "
-            "try again."));
+            "failed. This can happend due to missing privileges or bad device "
+            "state. Consider running with root privileges or remove, rescan "
+            "and reset the device."));
   }
 
   // Get IOMMU group of device
