### 2025-07-30 #8401 @Puneetha-Ramachandra
@danielskatz - From my perspective, the authors have address all my concerns except adding details about how this framework compares with other existing frameworks (tracked here - https://github.com/VILLASframework/node/issues/929). This is pending update from the authors.
