### 2025-08-05 #8401 @michmx
Hi @danielskatz. I am looking and the only remaining part to address from my side in this iteration are the comments in the paper. The authors are discussing them at https://github.com/VILLASframework/node/pull/930.
