### 2025-08-16 #8401 @Puneetha-Ramachandra
Hi @danielskatz, @stv0g,
I saw the draft on comparison with Helics and Mosaik. That check's off all items on my checklist. The paper looks good to me. I believe that concludes my review.
