### 2025-08-16 b070360
fix(paper): Address nit comments from reviewers
Signed-off-by: Steffen Vogel <post@steffenvogel.de>
