### 2025-08-18 50e0137
fix(paper): Mention energy/power system related use case
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
