### 2025-08-18 43fbe7d
fix(paper): Rephrase paragraph about integration of external tools
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
