+++ CMakeLists.txt
@@ -7,7 +7,7 @@
 cmake_minimum_required(VERSION 3.14)
 
 project(villas-node
-    VERSION 0.0.0
+    VERSION 1.0.0
     DESCRIPTION "VILLASnode"
     HOMEPAGE_URL "https://www.fein-aachen.org/projects/villas-node/"
     LANGUAGES C CXX
