### 2025-08-18 86ad223
fix(node-opal-orchestra): Use snake_case for options
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
