### 2025-08-18 d977806
fix(node-opal-orchestra): Use std::optional instead of negative value for absent indices
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
