### 2025-08-18 922ed11
fix(node-opal-orchestra): Add comment to global lock
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
