### 2025-08-18 ac4025c
fix(node-opal-orchestra): Use shared pointers for DDF items
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
