+++ lib/nodes/opal_orchestra/ddf.cpp
@@ -6,6 +6,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+#include <memory>
 extern "C" {
 #include <RTAPI.h>
 }
@@ -71,19 +72,19 @@ void BusItem::toXml(xmlNode *parent, bool withDefault) const {
   }
 }
 
-DataItem *BusItem::upsertItem(std::vector<std::string> pathComponents,
-                              bool &inserted) {
+std::shared_ptr<DataItem>
+BusItem::upsertItem(std::vector<std::string> pathComponents, bool &inserted) {
   auto &name = pathComponents.front();
 
   auto it = items.find(name);
   if (it == items.end()) { // No item with this name exists. Create a new one.
     if (pathComponents.size() == 1) { // No bus, just a signal.
-      auto *item = new DataItem(name);
+      auto item = std::make_shared<DataItem>(name);
       items.emplace(name, item);
       inserted = true;
       return item;
     } else {
-      auto *bus = new BusItem(name);
+      auto bus = std::make_shared<BusItem>(name);
       items.emplace(name, bus);
 
       pathComponents.erase(pathComponents.begin());
@@ -91,7 +92,7 @@ DataItem *BusItem::upsertItem(std::vector<std::string> pathComponents,
     }
   } else {
     if (pathComponents.size() == 1) {
-      auto *item = dynamic_cast<DataItem *>(it->second);
+      auto item = std::dynamic_pointer_cast<DataItem>(it->second);
       if (!item) {
         throw RuntimeError("Item with name '{}' is not a data item",
                            pathComponents.front());
@@ -100,7 +101,7 @@ DataItem *BusItem::upsertItem(std::vector<std::string> pathComponents,
       inserted = false;
       return item;
     } else {
-      auto *bus = dynamic_cast<BusItem *>(it->second);
+      auto bus = std::dynamic_pointer_cast<BusItem>(it->second);
       if (!bus) {
         throw RuntimeError("Item with name '{}' is not a bus",
                            pathComponents.front());
@@ -112,23 +113,24 @@ DataItem *BusItem::upsertItem(std::vector<std::string> pathComponents,
   }
 }
 
-DataItem *DataSet::upsertItem(const std::string &path, bool &inserted) {
+std::shared_ptr<DataItem> DataSet::upsertItem(const std::string &path,
+                                              bool &inserted) {
   return upsertItem(split(path), inserted);
 }
 
-DataItem *DataSet::upsertItem(std::vector<std::string> pathComponents,
-                              bool &inserted) {
+std::shared_ptr<DataItem>
+DataSet::upsertItem(std::vector<std::string> pathComponents, bool &inserted) {
   auto &name = pathComponents.front();
 
   auto it = items.find(name);
   if (it == items.end()) { // No item with this name exists. Create a new one.
     if (pathComponents.size() == 1) { // No bus, just a signal.
-      auto *item = new DataItem(name);
+      auto item = std::make_shared<DataItem>(name);
       items.emplace(name, item);
       inserted = true;
       return item;
     } else {
-      auto *bus = new BusItem(name);
+      auto bus = std::make_shared<BusItem>(name);
       items.emplace(name, bus);
 
       pathComponents.erase(pathComponents.begin());
@@ -136,7 +138,7 @@ DataItem *DataSet::upsertItem(std::vector<std::string> pathComponents,
     }
   } else {
     if (pathComponents.size() == 1) {
-      auto *item = dynamic_cast<DataItem *>(it->second);
+      auto item = std::dynamic_pointer_cast<DataItem>(it->second);
       if (!item) {
[…]
