### 2025-08-19 #8401 @michmx
Hi @danielskatz . In the most recent version of the paper my comments have been addressed. I am closing #938.

Also I am done with the Review checklist. I confirm the most recent version of the documentation addressses my comments.
