### 2025-08-19 0aabd70
Update paper.md
Signed-off-by: Daniel S. Katz <d.katz@ieee.org>
