### 2025-08-19 3e35e6d
Update paper.bib
Signed-off-by: Daniel S. Katz <d.katz@ieee.org>
