+++ paper/paper.bib
@@ -67,7 +67,7 @@
 
 @article{Mehlmann2023,
   author = {Gert Mehlmann and Uwe Kühnapfel and Felix Wege and Alexander Winkens and Christian Scheibe and Steffen Vogel},
-  title = {The Kopernikus ENSURE Co-Demonstration Platform},
+  title = {The {K}opernikus ENSURE Co-Demonstration Platform},
   journal = {IEEE Open Journal of Power Electronics},
   volume = {4},
   pages = {987--1002},
