### 2025-08-20 ef31765
fix(paper): Address comments from Niklas
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
