+++ tools/format-all.sh
@@ -9,5 +9,5 @@
 TOP_DIR=$(git rev-parse --show-toplevel)
 
 find ${TOP_DIR} \( -iname "*.cpp" -o -iname "*.hpp" -o -iname "*.h" -o -iname "*.c" \) -a \
-                -not \( -path "${TOP_DIR}/fpga/thirdparty/*" -o -path "${TOP_DIR}/build*/*" \) | \
+                -not \( -path "${TOP_DIR}/thirdparty/*" -o -path "${TOP_DIR}/build*/*" \) | \
     xargs clang-format --verbose -i
