+++ tools/cppcheck-supressions.txt
@@ -1,7 +1,7 @@
 # SPDX-FileCopyrightText: 2025, OPAL-RT Germany GmbH
 # SPDX-License-Identifier: Apache-2.0
 
-*:*/fpga/thirdparty/*
+*:*/thirdparty/*
 *:*/villas.pb-c.h
 
 noValidConfiguration
