### 2025-08-20 879af8f
Update paper.md
Signed-off-by: Daniel S. Katz <d.katz@ieee.org>
