+++ paper/paper.md
@@ -42,8 +42,8 @@ bibliography: paper.bib
 
 # Summary
 
-VILLASnode is a multi-protocol gateway, designed to facilitate real-time data exchange between various components of geographically distributed real-time experiments. Components can be testbeds, digital real-time simulators, software tools, and physical devices. It is designed for the co-simulation of power system and related energy applications.
-VILLASnode was originally designed for the co-simulation of electrical networks, but since then was further extended to cover a bigger variety of use cases and domains.
+VILLASnode is a multi-protocol gateway, designed to facilitate real-time data exchange between various components of geographically-distributed real-time experiments. Components can be testbeds, digital real-time simulators, software tools, and physical devices. It is designed for the co-simulation of power system and related energy applications.
+VILLASnode was originally designed for the co-simulation of electrical networks, but since then was further extended to cover a larger variety of use cases and domains.
 VILLASnode serves as the gateway that connects components across different infrastructures by providing a set of protocols and customized third-party implementations, e.g., different simulators. It enables seamless collaboration in research and testing environments while safeguarding the intellectual property of the infrastructures. The components of every infrastructure appear as a black box. The infrastructure does not need to share models or confidential information.
 
 VILLASnode is a set of Linux command line tools and shared library.
@@ -54,7 +54,7 @@ All components that are interfaced by the VILLASnode gateway are represented by 
 The basic data package, common for all node-types, includes timestamped data, constituting a sample. Up to 64 values can form a sample.
 Samples may need modification or filtering. VILLASnode supports hooks ($h$) for this purpose. Hooks are simple callback functions, which are called whenever a message is processed.
 Paths ($p$) take care of the processing and define the connections and dataflows between nodes.
-Node-types, hooks, and paths need to be initalized in a JSON configuration file which is passed when starting VILLASnode.
+Node-types, hooks, and paths need to be initalized in a JSON configuration file that is passed when starting VILLASnode.
 Figure 1 shows an example of an experiment where five different node-types are used, connected by three paths, using three hooks.
 It includes queues ($q$) and registers ($r$). Queues temporarily store data before data is forwarded to registers. Registers provide the possibility to (de-)mulitplex data and to create new samples.
 
@@ -65,26 +65,26 @@ Additionally, VILLASnode can be configured and controlled remotely by an HTTP RE
 VILLASnode is written in C++ and supports real-time execution on real-time Linux systems.
 To provide all necassary information, VILLASnode has detailed documentation [@villasnode_docs]. It includes installation recommendations and best practices for development as well as example configurations and beginners guides, so-called labs.
 
-Other open-source co-simulation tools are available, like Mosaik [@rohjans_mosaik_2013] or Helics [@hardy_helics_2024]. Mosaik uses a SimAPI to communicate with other simulators based on fixed timesteps. Integrated simulators need to support the SimAPI which requires extra implementations. Simulators cannot reuse existing supported protocols which VILLASnode makes use of and takes core of protocol conversion.
+VILLASnode can be compared with other available open-source co-simulation tools, such as Mosaik [@rohjans_mosaik_2013] or Helics [@hardy_helics_2024]. Mosaik uses a SimAPI to communicate with other simulators based on fixed timesteps. Integrated simulators need to support the SimAPI, which requires extra implementations. Simulators cannot reuse existing supported protocols, which VILLASnode makes use of and takes care of protocol conversion.
 Helics is designed for large-scale testing and supports bindings for different languages. It uses a broker that manages the communication between the simulators, whereas VILLASnode has a peer-to-peer architecture.
 
 # Statement of need
 
[…]
