### 2025-08-21 9273298
Bump version for JOSS publication
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
