### 2025-08-21 cb80e66
Bump version for JOSS publication
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
