### 2025-08-21 a5dc5bd
fix(compat): Add drop-in std::filesystem header for older systems
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
