+++ fpga/include/villas/fpga/utils.hpp
@@ -19,13 +19,13 @@ std::shared_ptr<fpga::Card> setupFpgaCard(const std::string &configFile,
                                           const std::string &fpgaName);
 
 std::shared_ptr<fpga::Card>
-createCard(json_t *config, const std::filesystem::path &searchPath,
+createCard(json_t *config, const fs::path &searchPath,
            std::shared_ptr<kernel::vfio::Container> vfioContainer,
            std::string card_name = "anonymous Card");
 
 int createCards(
     json_t *config, std::list<std::shared_ptr<fpga::Card>> &cards,
-    const std::filesystem::path &searchPath,
+    const fs::path &searchPath,
     std::shared_ptr<kernel::vfio::Container> vfioContainer = nullptr);
 int createCards(
     json_t *config, std::list<std::shared_ptr<fpga::Card>> &cards,
