+++ fpga/include/villas/fpga/pcie_card.hpp
@@ -10,7 +10,6 @@
 
 #pragma once
 
-#include <filesystem>
 #include <list>
 #include <set>
 #include <string>
@@ -20,6 +19,7 @@
 #include <villas/fpga/card.hpp>
 #include <villas/fpga/config.h>
 #include <villas/fpga/core.hpp>
+#include <villas/fs.hpp>
 #include <villas/kernel/devices/pci_device.hpp>
 #include <villas/kernel/vfio_container.hpp>
 #include <villas/memory.hpp>
@@ -64,8 +64,7 @@ class PCIeCardFactory : public plugin::Plugin {
 public:
   static std::shared_ptr<PCIeCard>
   make(json_t *json, std::string card_name,
-       std::shared_ptr<kernel::vfio::Container> vc,
-       const std::filesystem::path &searchPath);
+       std::shared_ptr<kernel::vfio::Container> vc, const fs::path &searchPath);
 
   static PCIeCard *make() { return new PCIeCard(); }
 
