+++ common/include/villas/utils.hpp
@@ -11,7 +11,6 @@
 #include <cassert>
 #include <cstdint>
 #include <cstdlib>
-#include <filesystem>
 #include <list>
 #include <string>
 #include <vector>
@@ -22,6 +21,7 @@
 #include <sys/types.h>
 
 #include <villas/config.hpp>
+#include <villas/fs.hpp>
 
 #ifdef __GNUC__
 #define LIKELY(x) __builtin_expect((x), 1)
@@ -212,9 +212,8 @@ template <class... Ts> struct overloaded : Ts... {
 // Explicit deduction guide (not needed as of C++20)
 template <class... Ts> overloaded(Ts...) -> overloaded<Ts...>;
 
-void write_to_file(std::string data, const std::filesystem::path file);
-std::vector<std::string>
-read_names_in_directory(const std::filesystem::path &directory);
+void write_to_file(std::string data, const fs::path file);
+std::vector<std::string> read_names_in_directory(const fs::path &directory);
 
 namespace base64 {
 
