+++ fpga/lib/node.cpp
@@ -5,8 +5,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include <map>
 #include <stdexcept>
+#include <unordered_map>
 
 #include <jansson.h>
 
