+++ lib/nodes/iec61850_goose.cpp
@@ -658,7 +658,7 @@ void GooseNode::parseSubscriber(json_t *json, GooseNode::SubscriberConfig &sc) {
 }
 
 void GooseNode::parseSubscribers(
-    json_t *json, std::map<std::string, InputEventContext> &ctx) {
+    json_t *json, std::unordered_map<std::string, InputEventContext> &ctx) {
   char const *key;
   json_t *json_subscriber;
 
