### 2025-08-21 #8401 @danielskatz
please see https://help.zenodo.org/docs/deposit/manage-records/
