### 2025-08-26 c7489c0
fix(node-opal-orchestra): Avoid passing const std::string references
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
