### 2025-08-26 a4e8a82
fix(node-opal-orchestra): Cleanup CMakeLists
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
