+++ CMakeLists.txt
@@ -189,8 +189,8 @@ cmake_dependent_option(WITH_NODE_EXAMPLE        "Build with example node-type"  
 cmake_dependent_option(WITH_NODE_EXEC           "Build with exec node-type"                             "${WITH_DEFAULTS}" "" OFF)
 cmake_dependent_option(WITH_NODE_FILE           "Build with file node-type"                             "${WITH_DEFAULTS}" "" OFF)
 cmake_dependent_option(WITH_NODE_FPGA           "Build with fpga node-type"                             "${WITH_DEFAULTS}" "WITH_FPGA" OFF)
-cmake_dependent_option(WITH_NODE_IEC61850       "Build with iec61850 node-types"                        "${WITH_DEFAULTS}" "LIBIEC61850_FOUND; NOT WITHOUT_GPL" OFF)
 cmake_dependent_option(WITH_NODE_IEC60870       "Build with iec60870 node-types"                        "${WITH_DEFAULTS}" "LIB60870_FOUND; NOT WITHOUT_GPL" OFF)
+cmake_dependent_option(WITH_NODE_IEC61850       "Build with iec61850 node-types"                        "${WITH_DEFAULTS}" "LIBIEC61850_FOUND; NOT WITHOUT_GPL" OFF)
 cmake_dependent_option(WITH_NODE_INFINIBAND     "Build with infiniband node-type"                       "${WITH_DEFAULTS}" "IBVerbs_FOUND; RDMACM_FOUND" OFF) # Infiniband node-type is currenly broken
 cmake_dependent_option(WITH_NODE_INFLUXDB       "Build with influxdb node-type"                         "${WITH_DEFAULTS}" "" OFF)
 cmake_dependent_option(WITH_NODE_KAFKA          "Build with kafka node-type"                            "${WITH_DEFAULTS}" "RDKAFKA_FOUND" OFF)
@@ -200,6 +200,7 @@ cmake_dependent_option(WITH_NODE_MQTT           "Build with mqtt node-type"     
 cmake_dependent_option(WITH_NODE_NANOMSG        "Build with nanomsg node-type"                          "${WITH_DEFAULTS}" "NANOMSG_FOUND" OFF)
 cmake_dependent_option(WITH_NODE_NGSI           "Build with ngsi node-type"                             "${WITH_DEFAULTS}" "" OFF)
 cmake_dependent_option(WITH_NODE_OPAL_ORCHESTRA "Build with the opal-orchestra node-type"               "${WITH_DEFAULTS}" "OpalOrchestra_FOUND; LibXml2_FOUND" OFF)
+cmake_dependent_option(WITH_NODE_OPENDSS        "Build with opendss node-type"                          "${WITH_DEFAULTS}" "OpenDSSC_FOUND" OFF)
 cmake_dependent_option(WITH_NODE_REDIS          "Build with redis node-type"                            "${WITH_DEFAULTS}" "HIREDIS_FOUND; REDISPP_FOUND" OFF)
 cmake_dependent_option(WITH_NODE_RTP            "Build with rtp node-type"                              "${WITH_DEFAULTS}" "re_FOUND" OFF)
 cmake_dependent_option(WITH_NODE_SHMEM          "Build with shmem node-type"                            "${WITH_DEFAULTS}" "HAS_SEMAPHORE; HAS_MMAN" OFF)
@@ -212,7 +213,6 @@ cmake_dependent_option(WITH_NODE_ULDAQ          "Build with uldaq node-type"    
 cmake_dependent_option(WITH_NODE_WEBRTC         "Build with webrtc node-type"                           "${WITH_DEFAULTS}" "WITH_WEB; LibDataChannel_FOUND" OFF)
[…]
