### 2025-08-26 f2b265d
fix(style): Format code with clang-format
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
