+++ common/include/villas/kernel/devices/linux_driver.hpp
@@ -33,11 +33,11 @@ private:
 public:
   LinuxDriver(const fs::path path)
       : LinuxDriver(path, path / fs::path(BIND_DEFAULT),
-                    path / fs::path(UNBIND_DEFAULT)) {};
+                    path / fs::path(UNBIND_DEFAULT)){};
 
   LinuxDriver(const fs::path path, const fs::path bind_path,
               const fs::path unbind_path)
-      : path(path), bind_path(bind_path), unbind_path(unbind_path) {};
+      : path(path), bind_path(bind_path), unbind_path(unbind_path){};
 
 public:
   void attach(const Device &device) const override;
