### 2025-08-26 186e880
fix(style): Format code with clang-format
Signed-off-by: Steffen Vogel <steffen.vogel@opal-rt.com>
